744-745 ХЫЫЫЫЫЫЫ)))

+++ printf/includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: kshanti <kshanti@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/21 15:08:11 by kshanti           #+#    #+#             */
-/*   Updated: 2020/12/01 18:16:02 by kshanti          ###   ########.fr       */
+/*   Updated: 2020/12/01 20:16:38 by kshanti          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,11 +31,11 @@ int				column_width(t_arg *tmp, long long num);
 int				column_precision(t_arg *tmp, long long num);
 int				column_width_str(t_arg *tmp, char *str);
 int				output_precision_str(t_arg *tmp, char *str);
-int             precision_poiner(t_arg *tmp, unsigned int n);
+int             precision_poiner(t_arg *tmp,long unsigned int n);
 void			out_to_16(long long n, char c);
-int       		pointer_length(unsigned int n);
+int       		pointer_length(long unsigned int n);
 void			ft_put_n_char(char ch, int n);
-int             width_pointer(t_arg *tmp, unsigned int n);
+int             width_pointer(t_arg *tmp,long unsigned int n);
 int				processor(t_arg *tmp, va_list *va);
 
 #endif
